chore: update spec + expose new methods
